+++ custom_components/spatial_presence/frontend/spatial-presence-card.js
@@ -866,7 +866,7 @@ function H(r, t) {
 function qt(r) {
   return typeof r == "string" && /^(\/|https?:\/\/)/i.test(r.trim());
 }
-const Re = "0.1.0-alpha.8", Wt = { states: {} }, J = class J extends z {
+const Re = "0.1.0-alpha.9", Wt = { states: {} }, J = class J extends z {
   constructor() {
     super(...arguments), this.hass = Wt, this.editorMode = !1, this._floorId = "", this._view = { x: 0, y: 0, width: 1200, height: 800 }, this._tool = "pan", this._draftPoints = [], this._showCoverage = !0, this._showTrails = !0, this._pointerMoved = !1, this._trails = /* @__PURE__ */ new Map(), this._fit = () => {
       const t = this._floor;